Working code using pyjq library
