+++ README.rst
@@ -51,6 +51,7 @@ FogLAMP is currently based on C/C++ and Python code. The packages needed to buil
 - *python3-pip*
 - *python-dbus*
 - *postgresql*
+- *autoconf*, *automake*, *build-essential*, *libtool*, *python-dev*
 
 On Ubuntu-based Linux distributions the packages can be installed with *apt-get*:
 ::
@@ -60,6 +61,7 @@ On Ubuntu-based Linux distributions the packages can be installed with *apt-get*
    apt-get install python3-pip
    apt-get install python-dbus
    apt-get install postgresql
+   apt-get install autoconf automake build-essential libtool python-dev
 
 You may need to use *sudo* to allow *apt-get* to install packages dependent upon your access rights.
